# Experiment with ArrowStream streaming