POC for fixing ArrowStream reading when source is not a local file.

Before:

```
ninja clickhouse-local && programs/clickhouse-local -q "SELECT count(), min(number), max(number) FROM url('http://127.0.0.1:3000/inf', "ArrowStream", 'number UInt64') FORMAT JSON" --logger.level trace
<Information> executeQuery: Read 100000000 rows, 762.94 MiB in 9.683175 sec., 10327191 rows/sec., 78.79 MiB/sec.
**<Debug> MemoryTracker: Peak memory usage (for query): 2.25 Gi**
```

After:

```
<Information> executeQuery: Read 100000000 rows, 762.94 MiB in 6.701079 sec., 14922969 rows/sec., 113.85 MiB/sec.
**<Debug> MemoryTracker: Peak memory usage (for query): 4.59 MiB.**
```

I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of reading from `ArrowStream` input format for sources other then local file (e.g. URL).