@KochetovNicolai 2021-04-12
```
zless clickhouse-server.log.gz | grep -a Fatal
2021.04.08 01:47:41.131832 [ 295 ] {} <Fatal> Application: Child process was terminated by signal 9 (KILL). If it is not done by 'forcestop' command or manually, the possible cause is OOM Killer (see 'dmesg' and look at the '/var/log/kern.log' for the details).
```